Avoid using Ember's internal symbol utility.
`symbol` is intended to give a good way to build an interned string (to
avoid the performance penalty of comparing strings stored as a rope).
Ember changed where `symbol` was kept upstream, breaking this addon.

In our usage scenarios, the thing we want to guarantee is that it is
difficult (not impossible since we export the value) to "forge" this
value. Using a function serves both purposes; it is fast for comparison,
and its identity cannot be forged (one `function() {} !== function()
{}`).