Makes sure we return the same pending promise if we try to load the same asset more than once
Without this change, when there is a second call to loadAsset for the same `uri` before the first one resolved, we return a resolved promise immediately. This is problematic since the consumer of loadAsset or loadBundle may think that it everything finished loading, for example, for the typical case with routes, we may decide to continue the transition.

This is another rare occurrence, but it can happen when a user clicks on a second link before we fully load the assets.