Fixup test assertions.
We cannot simply return the promise in the failure scenarios (since
returning a rejected promise will fail the test).  I have added
back a small assertion confirming the rejections in cases we _know_
should reject.

The CSS load failure scenario is a bit annoying/odd because we cannot
guarantee that we fail properly when the asset cannot be loaded (due
to lack of browser support for `onload`/`onerror` callbacks with the
`<link>` element).